fix(edition): move section detail overview into separate comp
