+++ src/app/views/edition-view/edition-outlets/edition-series-detail/edition-section-detail-overview/edition-section-detail-overview.component.html
@@ -1,0 +1,28 @@
+@if (
+    selectedSection &&
+    utils.isNotEmptyObject(selectedSection.complexTypes) &&
+    (utils.isNotEmptyArray(selectedSection.complexTypes.opus) ||
+        utils.isNotEmptyArray(selectedSection.complexTypes.mnr))
+) {
+    <div class="awg-edition-section-detail">
+        @if (utils.isNotEmptyArray(selectedSection.complexTypes.opus)) {
+            <div class="awg-edition-section-detail-opus para p-3 shadow border rounded-3">
+                <h5>nach Opusnummer:</h5>
+                <awg-edition-complex-card [complexes]="selectedSection.complexTypes.opus"></awg-edition-complex-card>
+            </div>
+        }
+        @if (utils.isNotEmptyArray(selectedSection.complexTypes.mnr)) {
+            <div class="awg-edition-section-detail-mnr p-3 shadow border rounded-3">
+                <h5>nach Moldenhauer-Nummer:</h5>
+                <awg-edition-complex-card [complexes]="selectedSection.complexTypes.mnr"></awg-edition-complex-card>
+            </div>
+        }
+    </div>
+} @else {
+    <div class="alert alert-info" role="alert">
+        <p class="text-muted" style="margin-bottom: 0">
+            [Diese Inhalte erscheinen im Zusammenhang der vollständigen Edition von AWG
+            {{ selectedSeries?.series.short }}/{{ selectedSection?.section.short }}.]
+        </p>
+    </div>
+}
